requirements.txt
diffusers==0.34.0
imageio==2.34.2
matplotlib==3.10.5
numpy==1.23.5
torch-fidelity==0.3.0